# Conflicts:
#	.gitignore
#	src/NetBox/CMakeLists.txt
#	src/NetBox/FarConfiguration.cpp
#	src/NetBox/FarConfiguration.h
#	src/NetBox/FarDialog.cpp
#	src/NetBox/FarDialog.h
#	src/NetBox/FarInterface.h
#	src/NetBox/FarPlugin.cpp
#	src/NetBox/FarPlugin.h
#	src/NetBox/FarPluginStrings.cpp
#	src/NetBox/FarUtils.h
#	src/NetBox/NetBox.cpp
#	src/NetBox/NetBox.rc
#	src/NetBox/NetBox.vcxproj
#	src/NetBox/NetBox.vs2015.vcxproj
#	src/NetBox/NetBoxEng.lng
#	src/NetBox/NetBoxRus.lng
#	src/NetBox/WinSCPDialogs.cpp
#	src/NetBox/WinSCPFileSystem.cpp
#	src/NetBox/WinSCPFileSystem.h
#	src/NetBox/WinSCPPlugin.cpp
#	src/NetBox/WinSCPPlugin.h
#	src/NetBox/XmlStorage.h
#	src/NetBox/plugin_version.hpp
#	src/base/Classes.hpp
#	src/base/MsgIDs.h
#	src/core/FtpFileSystem.cpp
#	src/core/HierarchicalStorage.h
#	src/core/ScpFileSystem.cpp
#	src/core/SessionInfo.cpp
#	src/core/SftpFileSystem.cpp
#	src/core/Terminal.cpp
#	src/core/Terminal.h
#	src/core/WebDAVFileSystem.cpp

+++ libs/putty/be_misc.c
@@ -36,7 +36,7 @@ void backend_socket_log(void *frontend, int type, SockAddr addr, int port,
 
             /* Suffix \r\n temporarily, so we can log to the terminal. */
             msg = dupprintf("%s\r\n", error_msg);
-            len = (int)strlen(msg);
+            len = strlen(msg);
             assert(len >= 2);
 
             log_to_term = conf_get_int(conf, CONF_proxy_log_to_term);
@@ -100,7 +100,7 @@ void log_proxy_stderr(Plug plug, bufchain *buf, const void *vdata, int len)
         /*
          * Advance past the newline.
          */
-        pos += (int)(nlpos+1 - (data + pos));
+        pos += nlpos+1 - (data + pos);
     }
 
     /*
